Create engine search.py
Using Python to ask a web page to run a search:

I wrote this script with some online help just to start developing my coding skills. The script supposed to return the data for searching an item in a particular website.  But it never brings a readable data back! I tried this code on other websites too but they are all give me the same results! Does anyone know what the problem is and where I am going wrong?